+++ pkg/apis/core/v1/helper/helpers.go
@@ -316,53 +316,6 @@ func AddOrUpdateTolerationInPod(pod *v1.Pod, toleration *v1.Toleration) bool {
 	return AddOrUpdateTolerationInPodSpec(&pod.Spec, toleration)
 }
 
-// TolerationsTolerateTaint checks if taint is tolerated by any of the tolerations.
-func TolerationsTolerateTaint(tolerations []v1.Toleration, taint *v1.Taint) bool {
-	for i := range tolerations {
-		if tolerations[i].ToleratesTaint(taint) {
-			return true
-		}
-	}
-	return false
-}
-
-type taintsFilterFunc func(*v1.Taint) bool
-
-// TolerationsTolerateTaintsWithFilter checks if given tolerations tolerates
-// all the taints that apply to the filter in given taint list.
-// DEPRECATED: Please use FindMatchingUntoleratedTaint instead.
-func TolerationsTolerateTaintsWithFilter(tolerations []v1.Toleration, taints []v1.Taint, applyFilter taintsFilterFunc) bool {
-	_, isUntolerated := FindMatchingUntoleratedTaint(taints, tolerations, applyFilter)
-	return !isUntolerated
-}
-
-// FindMatchingUntoleratedTaint checks if the given tolerations tolerates
-// all the filtered taints, and returns the first taint without a toleration
-func FindMatchingUntoleratedTaint(taints []v1.Taint, tolerations []v1.Toleration, inclusionFilter taintsFilterFunc) (v1.Taint, bool) {
-	filteredTaints := getFilteredTaints(taints, inclusionFilter)
-	for _, taint := range filteredTaints {
-		if !TolerationsTolerateTaint(tolerations, &taint) {
-			return taint, true
-		}
-	}
-	return v1.Taint{}, false
-}
-
-// getFilteredTaints returns a list of taints satisfying the filter predicate
-func getFilteredTaints(taints []v1.Taint, inclusionFilter taintsFilterFunc) []v1.Taint {
-	if inclusionFilter == nil {
-		return taints
-	}
-	filteredTaints := []v1.Taint{}
-	for _, taint := range taints {
-		if !inclusionFilter(&taint) {
-			continue
-		}
-		filteredTaints = append(filteredTaints, taint)
-	}
-	return filteredTaints
-}
-
 // GetMatchingTolerations returns true and list of Tolerations matching all Taints if all are tolerated, or false otherwise.
 func GetMatchingTolerations(taints []v1.Taint, tolerations []v1.Toleration) (bool, []v1.Toleration) {
 	if len(taints) == 0 {
