+++ test/integration/framework/BUILD
@@ -38,7 +38,6 @@ go_library(
         "//cmd/kube-apiserver/app:go_default_library",
         "//cmd/kube-apiserver/app/options:go_default_library",
         "//pkg/api/legacyscheme:go_default_library",
-        "//pkg/apis/core/v1/helper:go_default_library",
         "//pkg/controller/nodelifecycle:go_default_library",
         "//pkg/controlplane:go_default_library",
         "//pkg/generated/openapi:go_default_library",
@@ -73,6 +72,7 @@ go_library(
         "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//staging/src/k8s.io/client-go/util/cert:go_default_library",
         "//staging/src/k8s.io/component-base/version:go_default_library",
+        "//staging/src/k8s.io/component-helpers/scheduling/corev1:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/go-openapi/spec:go_default_library",
         "//vendor/github.com/google/uuid:go_default_library",
